webdav: Add XML and HTTP handler support for PROPPATCH.
This change adds support to parse and handle PROPPATCH requests. It adds the
Patch method to the PropSystem interface. In order to keep this CL small,
it does not add support to set "dead" DAV properties. Instead, any patched
property is reported with HTTP status code 403 Forbidden. Once this CL is
accepted, I will submit a CL to store dead DAV properties in-memory. The
litmus test coverage of the 'props' test suite remains as-is:

16 tests were skipped, 14 tests run. 10 passed, 4 failed. 71.4%

Change-Id: I14a25464e94b3316c16976f79b4457bf646d0d24
Reviewed-on: https://go-review.googlesource.com/8937
Reviewed-by: Nigel Tao <nigeltao@golang.org>