http2: fix bug where '*' as a valid :path value in Transport
In the most recent commit (git 3a1f9e, CL 27632) I introduced a bug
where I improperly started rejecting an outgoing :path of '*'.  That
CL was never vendored into std. Once I tried to vendor it and run the
tests, the tests failed due to this regression.

So, fix the bug and allow '*' (e.g. for OPTIONS requests). No test
updates because the tests are already in std, and now pass (again).

Change-Id: Ibb91a17145d381da0fd0ea193e8a1bd59f03c3fa
Reviewed-on: https://go-review.googlesource.com/29070
Reviewed-by: Chris Broadfoot <cbro@golang.org>