publicsuffix: allow to encode up to 1023 children in the table
The number of children (494) got pretty close to the available maximum
number of children (511) in the table. This number increased over the
last year by roughly 65 indicating that we would run out of space in
the next three months. The other fields still have enough room left.
The following table show the growth over the last 1.5 years:

    # Commit Date Children TextOffset Length High Low
    b05061f 2017-09-18 494 28750 36 8407 8402
    8ba35e8 2017-09-06 494 28750 36 8407 8402
    98e6c6e 2017-06-14 479 28411 36 8262 8257
    eb5e097 2017-01-26 466 28023 36 8121 8120
    28ba0a7 2016-10-20 434 27930 36 8135 8134
    dff4834 2016-08-11 424 27866 36 8062 8051
    b27498f 2016-07-07 421 27811 36 8049 8038
    2d1be0a 2016-06-09 417 27680 36 8029 8018
    2b3c342 2016-03-04 409 27059 36 7887 7886
    0d6fcde 2016-02-01 406 26999 36 7868 7867
    b06e34d 2016-01-20 405 26986 36 7863 7862

Given this rate of grow of max text offset it will overflow in 2021.
Thus use the last of the available 32 bits to encode more children.

Change-Id: I04db02100b202f220a0b4ee509f868db031fd8ab
Reviewed-on: https://go-review.googlesource.com/64330
Reviewed-by: Nigel Tao <nigeltao@golang.org>